feat: add Actions and Expanders

+++ CodexNaturalis/src/main/java/it/polimi/ingsw/model/playerReleted/AgentRelated/ActionRelated/Action.java
@@ -4,11 +4,9 @@ import it.polimi.ingsw.controller.GameController;
 import it.polimi.ingsw.model.playerReleted.AgentRelated.Node;
 import it.polimi.ingsw.model.playerReleted.AgentRelated.StateRelated.State;
 
-import java.util.function.Predicate;
-
 public interface Action {
     void actOnState(State state);
     void actOnGame(GameController controller, String nickname);
-    int getSelectionScore(Node node);
+    float getSelectionScore(Node node);
     int simulate(State state);
 }
